Add indexing config modal, search mode checkboxes, and multi-path retrieval
Replace the inline index panel with a clickable index badge and a modal
for indexing configuration (schema source, JSON editor, embeddings toggle,
progress view, post-indexing summary).

Replace exclusive search mode radio buttons with additive checkboxes:
Agentic is always on as the base exploration mode, with Semantic and
Metadata as optional retrieval add-ons. The agent always has access to
all tools — indexed retrieval provides a fast initial pass before
filesystem deep dives.

Backend changes:
- search/query.py: enable_semantic/enable_metadata params on search()
- agent.py: search flags state, dynamic system prompt builder
- workflow.py: pass retrieval flags through InputEvent/WorkflowState
- server.py: read flags from WebSocket, enhanced index_status with schema_fields

Benchmarked on 11-doc corpus: Agentic+Semantic+Metadata is ~22% faster
than pure Agentic exploration (16.7s vs 21.4s avg).

+++ src/fs_explorer/search/ranker.py
@@ -16,7 +16,7 @@ class RankedDocument:
     absolute_path: str
     position: int | None
     text: str
-    semantic_score: int
+    semantic_score: float
     metadata_score: int
 
     @property
@@ -34,7 +34,9 @@ class RankedDocument:
         return "metadata"
 
 
-def rank_documents(documents: list[RankedDocument], *, limit: int) -> list[RankedDocument]:
+def rank_documents(
+    documents: list[RankedDocument], *, limit: int
+) -> list[RankedDocument]:
     """Sort merged retrieval results and apply limit."""
     ordered = sorted(
         documents,
